# docs/api/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Agent Banking Platform API
  description: Aggregated API from all microservices
  contact:
    name: Agent Banking Platform Team
servers:
- description: Gateway
  url: http://localhost:8080
- description: Production
  url: https://api.agentbanking.example.com
- description: Sandbox
  url: https://sandbox.agentbanking.example.com
paths:
...
  /api/v1/auth/me:
    get:
      tags:
      - auth-controller (auth-iam-service)
      operationId: getMyProfile
      security:
        - bearerAuth: []
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user. For external agents, includes linked agentId.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/token:
    post:
      tags:
      - auth-controller (auth-iam-service)
      operationId: authenticateUser
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags:
      - auth-controller (auth-iam-service)
      operationId: refreshToken
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/revoke:
    post:
      tags:
      - auth-controller (auth-iam-service)
      operationId: revokeToken
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeTokenRequest'
        required: true
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/password/reset:
    post:
      tags:
      - user-management-controller (auth-iam-service)
      operationId: resetPassword
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetPasswordResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/backoffice/audit-logs:
    get:
      tags:
      - audit-log-controller (onboarding-service)
      operationId: getAuditLogs
      security:
        - bearerAuth: []
      parameters:
      - name: entityType
        in: query
        required: false
        schema:
          type: string
      - name: fromDate
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: toDate
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      required:
      - status
      - error
      type: object
      properties:
        status:
          type: string
          enum: [FAILED]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            action_code:
              type: string
              enum: [DECLINE, RETRY, REVIEW]
            trace_id:
              type: string
            timestamp:
              type: string
              format: date-time

    # ============================================================
    # TRANSACTION ORCHESTRATOR SCHEMAS (NEW - 2026-04-05)
    # ============================================================

    # Transaction Type Enum
...
    TokenRequest:
      required:
      - username
      - password
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        grantType:
          type: string
          enum: [password, refresh_token]
          default: password
          minLength: 8
          maxLength: 64
    TokenResponse:
      required:
      - access_token
      - refresh_token
      - expires_in
      - token_type
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int32
          description: Token expiry in seconds
        token_type:
          type: string
          enum: [Bearer]
    MyProfileResponse:
      type: object
      properties:
        userId:
          type: string
          nullable: true
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        userType:
          type: string
          enum: [INTERNAL, EXTERNAL]
        status:
          type: string
        agentId:
          type: string
          nullable: true
          description: The linked agent ID for external users
        mustChangePassword:
          type: boolean
        permissions:
          type: array
          items:
            type: string
    RefreshTokenRequest:
      required:
      - refresh_token
      type: object
      properties:
        refresh_token:
          type: string
    RevokeTokenRequest:
      type: object
      properties:
        token:
          type: string
          description: Optional. If not provided, revokes all sessions for the current user.

    # Password Management
...
    ResetPasswordRequest:
      required:
      - username
      - otp
      - newPassword
      type: object
      properties:
        username:
          type: string
        otp:
          type: string
          minLength: 6
          maxLength: 6
        newPassword:
          type: string
          minLength: 8
          maxLength: 64
    ResetPasswordResponse:
      type: object
      properties:
        message:
          type: string
...